ci: use single dynamic build-test action
Add a GitHub action that dynamically generates the matrix of Dockerfiles
to test based on what files have changed. This avoids having to have
generated workflows for each version x variant combo.

It looks for any changed Dockerfiles, or docker-entrypoint.sh. If any of
the testing files are changed, it will test all Dockerfiles