fix: docker-entrypoint.sh file handling, closes #1456
The docker-entrypoint.sh script added in
https://github.com/nodejs/docker-node/issues/1039 is intended to run the
supplied command with "node" if it contains a "-" or doesn't correspond
to a system command. In Alpine, this doesn't work if the supplied
command corresponds to a regular, non-executable JS file.

The root issue is a bug in ash/dash: its implementation of "command -v"
incorrectly outputs the supplied command_name even for non-executable
files. This is a violation of the POSIX standard and has been reported
to the Debian team in
https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=874264, though there's
been no activity in several years.

As a workaround, this adds an additional check to docker-entrypoint.sh
for regular files that aren't marked as executable.