chore: Simplify yarn and alpine version update
Before commit 3f6cb089 (fix: only update yarn if there are other changes
to the dockerfile), yarnVersion depended on SKIP, so the substitute that
followed made sense. This commit removed the condition and changed it to
just substitute the version with itself, which doesn't make much sense.

Just remove it, and also move the real replace inside SKIP condition.

Also remove initial assignment to alpine_version, which is never used.